[feature/validation-wrapping] Rename ValidationHandler and update references
Moved the ValidationHandler class to the Validation namespace and updated all relevant references in tests and documentation. This refactor improves the project structure by grouping validation-related code logically.
